+++ src/primitives/sphere.rs
@@ -4,11 +4,11 @@ use super::SignedDistance;
 
 pub struct Sphere {
     position: Vector3,
-    radius: f32,
+    radius: f64,
 }
 
 impl Sphere {
-    pub fn new(position: impl Into<Vector3>, radius: f32) -> Self {
+    pub fn new(position: impl Into<Vector3>, radius: f64) -> Self {
         Sphere {
             position: position.into(),
             radius,
@@ -17,7 +17,7 @@ impl Sphere {
 }
 
 impl SignedDistance for Sphere {
-    fn distance_from(&self, position: Vector3) -> f32 {
+    fn distance_from(&self, position: Vector3) -> f64 {
         (position - self.position).magnitude() - self.radius
     }
 }
